Refactor of bpf programs to easier integrate new ones
