+++ bm-runner/monitors/bpftrace_programs/__init__.py
@@ -1,0 +1,1 @@
+__all__ = [ "bpf_block_req", "bpf_cgroup_attach_task", "bpf_cgroup_charge", "bpf_cgroup_fork", "bpf_cgroup_procs_write", "bpf_cgroup_rstat_flush", "bpf_cgroup_rstat_lock_cont", "bpf_copy_from_user", "bpf_copy_to_user", "bpf_ext4_mb_new_blocks", "bpf_flush_tlb_mm_range", "bpf_irq_handler", "bpf_mutex_lock", "bpf_native_queued_spin_lock_slowpath", "bpf_page_fault_io", "bpf_page_fault_kernel", "bpf_page_fault_kvm", "bpf_page_fault_user", "bpf_queued_read_lock_slowpath", "bpf_queued_write_lock_slowpath", "bpf_raw_spin_lock", "bpf_rwsem_down_read_slowpath", "bpf_rwsem_down_write_slowpath", "bpf_sched_domains_mutex", "bpf_sched_fork", "bpf_sched_latency", "bpf_sched_migrate_task", "bpf_vfs_read_latency"]
