+++ bm-runner/monitors/bpftrace_programs/bpf_cgroup_attach_task.py
@@ -1,11 +1,12 @@
 # Copyright (C) Huawei Technologies Co., Ltd. 2026. All rights reserved.
 # SPDX-License-Identifier: MIT
 
-import os
-import pandas as pd
-from monitors.bpftrace_programs.bpf_program import BPFProgram
+from monitors.bpf_program import BPFProgram
+from monitors.bpf_parser_histograms import BPFParserHistograms
 
 class BPFCGroupAttachTask(BPFProgram):
+    name = "cgroup_attach_task"
+    parser = BPFParserHistograms
     program = """
 kprobe:cgroup_attach_task
 / __FILTER_CPU__ && __FILTER_PID__ /
@@ -14,18 +15,7 @@ kprobe:cgroup_attach_task
 }
 kretprobe:cgroup_attach_task /@start[tid]/ {
     $duration = (nsecs - @start[tid]) / 1000;
-    @attach_latency = hist($duration);
+    @attach_latency[pid] = hist($duration);
     delete(@start[tid]);
 }
 """
-
-    filename = "bpf_cgroup_attach_task.log"
-    csv_key = "cgroup_attach_task"
-
-    def collect_results(self, output_dir: str, PIDs: list[int]) -> str:
-        result = ""
-        filepath = os.path.join(output_dir, self.filename)
-        df = self.parse_histogram(filepath)
-        result += self.results_histogram_min_max_avg(df=df)
-        result += self.results_histogram_histogram(df=df)
-        return result
