+++ bm-runner/monitors/bpftrace_programs/bpf_cgroup_procs_write.py
@@ -3,9 +3,12 @@
 
 import os
 import pandas as pd
-from monitors.bpftrace_programs.bpf_program import BPFProgram
+from monitors.bpf_program import BPFProgram
+from monitors.bpf_parser_histograms import BPFParserHistograms
 
 class BPFCGroupProcsWrite(BPFProgram):
+    name = "cgroup_procs_write"
+    parser = BPFParserHistograms()
     program = """
 kprobe:__cgroup_procs_write 
 / __FILTER_CPU__ && __FILTER_PID__ /
@@ -14,17 +17,7 @@ kprobe:__cgroup_procs_write
 }
 kretprobe:__cgroup_procs_write /@start[tid]/ {
     $duration = (nsecs - @start[tid]) / 1000;
-    @migration_latency = hist($duration);
+    @migration_latency[pid] = hist($duration);
     delete(@start[tid]);
 }
 """
-    filename = "bpf_cgroup_procs_write.log"
-    csv_key = "cgroup_procs_write"
-
-    def collect_results(self, output_dir: str, PIDs: list[int]) -> str:
-        result = ""
-        filepath = os.path.join(output_dir, self.filename)
-        df = self.parse_histogram(filepath)
-        result += self.results_histogram_min_max_avg(df=df)
-        result += self.results_histogram_histogram(df=df)
-        return result
