+++ bm-runner/monitors/bpftrace_programs/bpf_cgroup_rstat_lock.py
@@ -4,11 +4,14 @@
 from monitors.bpf_program import BPFProgram
 from monitors.bpf_parser_counts import BPFParserCounts
 
-class BPFSchedFork(BPFProgram):
-    name = "sched_fork"
+class BPFCGroupRstatLockCont(BPFProgram):
+    name = "cgroup_rstat_lock"
     parser = BPFParserCounts()
     program = """
-tracepoint:sched:sched_process_fork
+
+tracepoint:cgroup:cgroup_rstat_lock_contended
 / __FILTER_CPU__ && __FILTER_PID__ /
-{ @[args->parent_pid] = count(); }
+{
+    @contended[pid] = count();
+}
 """
