+++ bm-runner/monitors/bpftrace_programs/bpf_copy_from_user.py
@@ -1,11 +1,12 @@
 # Copyright (C) Huawei Technologies Co., Ltd. 2026. All rights reserved.
 # SPDX-License-Identifier: MIT
 
-import os
-import pandas as pd
-from monitors.bpftrace_programs.bpf_program import BPFProgram
+from monitors.bpf_program import BPFProgram
+from monitors.bpf_parser_histograms import BPFParserHistograms
 
 class BPFCopyFromUser(BPFProgram):
+    name = "copy_from_user"
+    parser = BPFParserHistograms()
     program = """
 kprobe:bpf_copy_from_user 
 / __FILTER_CPU__ && __FILTER_PID__ /
@@ -15,12 +16,3 @@ kretprobe:bpf_copy_from_user
 / @start[pid] /
 { @ns[pid] = hist(nsecs - @start[pid]); delete(@start[pid]); }
 """
-    csv_key = "copy_from_user"
-    filename = f"bpf_{csv_key}.log"
-
-    def collect_results(self, output_dir: str, PIDs: list[int]) -> str:
-        result = ""
-        filepath = os.path.join(output_dir, self.filename)
-        df = self.parse_histograms(filepath)
-        result += self.results_histograms_histogram(df=df, PIDs = PIDs)
-        return result
