+++ bm-runner/monitors/bpftrace_programs/bpf_ext4_mb_new_blocks.py
@@ -1,11 +1,11 @@
 # Copyright (C) Huawei Technologies Co., Ltd. 2026. All rights reserved.
 # SPDX-License-Identifier: MIT
 
-import os
-import pandas as pd
-from monitors.bpftrace_programs.bpf_program import BPFProgram
+from monitors.bpf_program import BPFProgram
+from monitors.bpf_parser_histograms import BPFParserHistograms
 
 class BPFExt4MbNewBLocks(BPFProgram):
+    name = "ext4_mb_new_blocks"
     program = """
 kprobe:ext4_mb_new_blocks 
 / __FILTER_CPU__ && __FILTER_PID__ /
@@ -15,12 +15,3 @@ kretprobe:ext4_mb_new_blocks
 / @start[pid] /
 { @ns[pid] = hist(nsecs - @start[pid]); delete(@start[pid]); }
 """
-    csv_key = "ext4_mb_new_blocks"
-    filename = f"bpf_{csv_key}.log"
-
-    def collect_results(self, output_dir: str, PIDs: list[int]) -> str:
-        result = ""
-        filepath = os.path.join(output_dir, self.filename)
-        df = self.parse_histograms(filepath)
-        result += self.results_histograms_histogram(df=df, PIDs = PIDs)
-        return result
