+++ bm-runner/monitors/bpftrace_programs/bpf_mutex_lock.py
@@ -3,15 +3,16 @@
 
 from monitors.bpf_program import BPFProgram
 from monitors.bpf_parser_histograms import BPFParserHistograms
-class BPFIrqHandler(BPFProgram):
-    name = "irq_handler"
+
+class BPFMutexLock(BPFProgram):
+    name = "mutex_lock"
     parser = BPFParserHistograms()
     program = """
-tracepoint:irq:irq_handler_entry
+kprobe:mutex_lock
 / __FILTER_CPU__ && __FILTER_PID__ /
-{ @start[pid] = nsecs; }
+{ @start[tid] = nsecs; }
 
-tracepoint:irq:irq_handler_exit
-/ @start[pid] /
-{ @ns[pid] = hist(nsecs - @start[pid]); delete(@start[pid]); }
+kprobe:mutex_unlock
+/ @start[tid] /
+{ @ns[pid] = hist(nsecs - @start[tid]); delete(@start[tid]); }
 """
