+++ bm-runner/monitors/bpftrace_programs/bpf_page_fault_user.py
@@ -4,11 +4,11 @@
 from monitors.bpf_program import BPFProgram
 from monitors.bpf_parser_counts import BPFParserCounts
 
-class BPFSchedFork(BPFProgram):
-    name = "sched_fork"
+class BPFPageFaultUser(BPFProgram):
+    name = "page_fault_user"
     parser = BPFParserCounts()
     program = """
-tracepoint:sched:sched_process_fork
+tracepoint:exceptions:page_fault_user
 / __FILTER_CPU__ && __FILTER_PID__ /
-{ @[args->parent_pid] = count(); }
+{ @fault[pid] = count(); }
 """
