+++ bm-runner/monitors/bpftrace_programs/bpf_queued_read_lock_slowpath.py
@@ -3,15 +3,16 @@
 
 from monitors.bpf_program import BPFProgram
 from monitors.bpf_parser_histograms import BPFParserHistograms
-class BPFIrqHandler(BPFProgram):
-    name = "irq_handler"
+
+class BPFQueuedReadLockSlowpath(BPFProgram):
+    name = "queued_read_lock_slowpath"
     parser = BPFParserHistograms()
     program = """
-tracepoint:irq:irq_handler_entry
+kprobe:queued_read_lock_slowpath
 / __FILTER_CPU__ && __FILTER_PID__ /
-{ @start[pid] = nsecs; }
+{ @start[tid] = nsecs; }
 
-tracepoint:irq:irq_handler_exit
-/ @start[pid] /
-{ @ns[pid] = hist(nsecs - @start[pid]); delete(@start[pid]); }
+kretprobe:queued_read_lock_slowpath
+/ @start[tid] /
+{ @ns[pid] = hist(nsecs - @start[tid]); delete(@start[tid]); }
 """
